Fix
Виправлено зауваження
