+++ models/contact.js
@@ -1,15 +1,19 @@
 import { Schema, model } from "mongoose";
 
+
 const contsctSchema = new Schema({
     name: {
       type: String,
+      match: /^[a-zA-Zа-яА-Я]+(([' -][a-zA-Zа-яА-Я ])?[a-zA-Zа-яА-Я]*)*$/,
       required: [true, 'Set name for contact'],
     },
     email: {
       type: String,
+      match: /.+\@.+\..+/,
     },
     phone: {
       type: String,
+      match: /^((\+)?(3)?(8)?[- ]?)?(\(?\d{3}\)?[- ]?)?\d{3}[- ]?\d{2}[- ]?\d{2}$/,
     },
     favorite: {
       type: Boolean,
@@ -17,11 +21,6 @@ const contsctSchema = new Schema({
     },
   }, {versionKey: false, timestamps: true});
 
-
-// contsctSchema.post("save", (error, data, next) => {
-//   error.status = 400;
-//   next();
-// });  ----- если не прошло валидацию? пример с датой релиза.
 
 const Contact = model("contact", contsctSchema);
