+++ routes/api/contacts-router.js
@@ -1,29 +1,10 @@
 import express from "express";
-import Joi from "joi";
 
 import contactsService from "../../models/contacts.js";
 import { HttpError } from '../../helpers/index.js';
-import {isEmptyBody, isValidId } from '../../middlewars/index.js';
+import { isEmptyBody, isValidId } from '../../middlewars/index.js';
+import contactsSchemas from '../../shema/contacts-schemas.js';
 
-
-
-const contactsAddSchema = Joi.object({
-  name: Joi.string().required().messages({
-    "any.required": `"name" must be exist`,
-  }),
-  email: Joi.string().required().messages({
-    "any.required": `"email" must be exist`,
-  }),
-  phone: Joi.string().required().messages({
-    "any.required": `"phone" must be exist`,
-  }),
-  favorite: Joi.boolean(),
-});
-
-
-const contactsUpdateFavoriteShema = Joi.object({
-  favorite: Joi.boolean().required()
-});
 
 
 const validateBody = schema => {
@@ -44,12 +25,12 @@ contactsRouter.get('/', contactsService.listContacts);
 
 contactsRouter.get('/:id', isValidId, contactsService.getContactById)
 
-contactsRouter.post('/', isEmptyBody, validateBody(contactsAddSchema), contactsService.addContact);
+contactsRouter.post('/', isEmptyBody, validateBody(contactsSchemas.contactsAddSchema), contactsService.addContact);
 
 contactsRouter.delete('/:id', isValidId, contactsService.removeContact);
 
-contactsRouter.put('/:id', isValidId, isEmptyBody, validateBody(contactsAddSchema), contactsService.updateContact);
+contactsRouter.put('/:id', isValidId, isEmptyBody, validateBody(contactsSchemas.contactsAddSchema), contactsService.updateContact);
 
-contactsRouter.patch('/:id/favorite', isValidId, isEmptyBody, validateBody(contactsUpdateFavoriteShema), contactsService.updateStatusContact);
+contactsRouter.patch('/:id/favorite', isValidId, isEmptyBody, validateBody(contactsSchemas.contactsUpdateFavoriteShema), contactsService.updateStatusContact);
 
 export default contactsRouter;
