+++ shema/contacts-schemas.js
@@ -1,0 +1,26 @@
+import Joi from "joi";
+
+const contactsAddSchema = Joi.object({
+  name: Joi.string().required().messages({
+    "any.required": `"name" must be exist`,
+  }),
+  email: Joi.string().required().messages({
+    "any.required": `"email" must be exist`,
+  }),
+  phone: Joi.string().required().messages({
+    "any.required": `"phone" must be exist`,
+  }),
+  favorite: Joi.boolean(),
+});
+
+
+
+const contactsUpdateFavoriteShema = Joi.object({
+  favorite: Joi.boolean().required()
+});
+
+
+export default {
+    contactsAddSchema,
+    contactsUpdateFavoriteShema,
+}
